refactor: extract registerServiceWroker and removeSplashScreen to their own modules
